# docs/open-api-spec/swagger.yaml
definitions:
...
  controllers.HyperlinkedDeliveryJobModel:
    properties:
      channelURL:
        type: string
      consumerURL:
        type: string
      jobRequeueURL:
        type: string
      listenerEndpoint:
        type: string
      listenerName:
        type: string
      messageURL:
        type: string
      producerURL:
        type: string
      status:
        type: string
      statusChangedAt:
        type: string
    type: object
  controllers.JobListResult:
    properties:
      result:
        items:
          $ref: '#/definitions/controllers.QueuedDeliveryJobModel'
        type: array
    type: object
  controllers.JobStatusUpdateRequest:
    properties:
      IncrementalTimeout:
        type: integer
      NextState:
        type: string
    type: object
...
  controllers.QeuedMessageModel:
    properties:
      contentType:
        type: string
      headers:
        $ref: '#/definitions/data.HeadersMap'
      messageID:
        type: string
      payload:
        type: string
    type: object
  controllers.QueuedDeliveryJobModel:
    properties:
      id:
        items:
          format: int32
          type: integer
        type: array
      incrementalTimeout:
        description: in second
        type: integer
      message:
        $ref: '#/definitions/controllers.QeuedMessageModel'
      priority:
        type: integer
    type: object
...
  data.HeadersMap:
    additionalProperties:
      type: string
    type: object
info:
  contact: {}
  description: This is a fully HTTP-based Pub/Sub Broker designed to simplify system
    architecture in SOA or Microservice architecture.
  title: Webhook Broker API
  version: 1.0.0
paths:
...
  /channel/{channelId}/consumer/{consumerId}/job/{jobId}:
    delete:
      description: Delete a single dead job.
      parameters:
      - description: Channel ID
        in: path
        name: channelId
        required: true
        type: string
      - description: Consumer ID
        in: path
        name: consumerId
        required: true
        type: string
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: string
      - description: Channel authentication token
        in: header
        name: X-Broker-Channel-Token
        required: true
        type: string
      - description: Consumer authentication token
        in: header
        name: X-Broker-Consumer-Token
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - ChannelToken: []
        ConsumerToken: []
      summary: Delete Single Dead Job
      tags:
      - DLQ
    get:
      description: Retrieves detailed information about a specific delivery job.
      parameters:
      - description: Channel ID
        in: path
        name: channelId
        required: true
        type: string
      - description: Consumer ID
        in: path
        name: consumerId
        required: true
        type: string
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: string
      - description: Channel authentication token
        in: header
        name: X-Broker-Channel-Token
        required: true
        type: string
      - description: Consumer authentication token
        in: header
        name: X-Broker-Consumer-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.HyperlinkedDeliveryJobModel'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "412":
          description: Precondition Failed
        "500":
          description: Internal Server Error
      security:
      - ChannelToken: []
        ConsumerToken: []
      summary: Get Job Details
      tags:
      - Jobs
    post:
      consumes:
      - application/json
      description: Updates the status of a delivery job. Used by pull-based consumers
        to transition jobs through their lifecycle.
      parameters:
      - description: Channel ID
        in: path
        name: channelId
        required: true
        type: string
      - description: Consumer ID
        in: path
        name: consumerId
        required: true
        type: string
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: string
      - description: Channel authentication token
        in: header
        name: X-Broker-Channel-Token
        required: true
        type: string
      - description: Consumer authentication token
        in: header
        name: X-Broker-Consumer-Token
        required: true
        type: string
      - description: Job update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.JobStatusUpdateRequest'
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "412":
          description: Precondition Failed
        "500":
          description: Internal Server Error
      security:
      - ChannelToken: []
        ConsumerToken: []
      summary: Update Job Status (Pull-Based)
      tags:
      - Jobs
  /channel/{channelId}/consumer/{consumerId}/job/{jobId}/requeue-dead-job:
    post:
      consumes:
      - application/json
      description: Requeues a specific dead job for another delivery attempt.
      parameters:
      - description: Channel ID
        in: path
        name: channelId
        required: true
        type: string
      - description: Consumer ID
        in: path
        name: consumerId
        required: true
        type: string
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: string
      - description: Channel authentication token
        in: header
        name: X-Broker-Channel-Token
        required: true
        type: string
      - description: Consumer authentication token
        in: header
        name: X-Broker-Consumer-Token
        required: true
        type: string
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "412":
          description: Precondition Failed
        "500":
          description: Internal Server Error
      security:
      - ChannelToken: []
        ConsumerToken: []
      summary: Requeue Single Dead Job
      tags:
      - Jobs
  /channel/{channelId}/consumer/{consumerId}/queued-jobs:
    get:
      description: Retrieves prioritized queued jobs for a pull-based consumer.
      parameters:
      - description: Channel ID
        in: path
        name: channelId
        required: true
        type: string
      - description: Consumer ID
        in: path
        name: consumerId
        required: true
        type: string
      - default: 25
        description: Page size
        in: query
        maximum: 100
        name: limit
        type: integer
      - description: Channel authentication token
        in: header
        name: X-Broker-Channel-Token
        required: true
        type: string
      - description: Consumer authentication token
        in: header
        name: X-Broker-Consumer-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.JobListResult'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "412":
          description: Precondition Failed
        "500":
          description: Internal Server Error
      security:
      - ChannelToken: []
        ConsumerToken: []
      summary: Get Queued Jobs for Consumer (Pull-Based)
      tags:
      - Jobs
...
securityDefinitions:
  ChannelToken:
    description: Channel authentication token issued when the channel was created
      or last rotated.
    in: header
    name: X-Broker-Channel-Token
    type: apiKey
  ConsumerToken:
    description: Consumer authentication token issued when the consumer was created
      or last rotated.
    in: header
    name: X-Broker-Consumer-Token
    type: apiKey
  ProducerID:
    description: Identifier of the producer originating the request.
    in: header
    name: X-Broker-Producer-ID
    type: apiKey
  ProducerToken:
    description: Producer authentication token issued when the producer was created
      or last rotated.
    in: header
    name: X-Broker-Producer-Token
    type: apiKey
swagger: "2.0"